Migrate structures::jboot to zerocopy

+++ src/extractors/jboot.rs
@@ -55,7 +55,7 @@ pub fn extract_jboot_sch2_kernel(
 
             // Validate the kernel data checksum
             if let Some(kernel_data) = file_data.get(kernel_start..kernel_end)
-                && crc32(kernel_data) == (sch2_header.kernel_checksum as u32)
+                && crc32(kernel_data) == sch2_header.kernel_checksum
             {
                 // Everything checks out ok
                 result.size = Some(sch2_header.header_size + sch2_header.kernel_size);
